Backend/requirements.txt
tensorflow
numpy
pillow
logging 